rename: RepairBanner 이름 RequestBanner로 변경
