+++ src/templates/repair/RepairHomeTemplate.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import PageHeader from '../../components/common/PageHeader';
 
 export default function RepairHomeTemplate() {
   return (
     <>
+      <PageHeader leftComponent={'뚝딱'} />
       <h1>뚝딱뚜깍</h1>
     </>
   );
